requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
langgraph>=0.0.26
langchain-google-genai>=0.0.6
langchain-groq>=0.1.0
groq>=0.4.0
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.11.0
pydantic>=2.5.0
structlog>=24.1.0
prometheus_client>=0.19.0
tenacity>=8.2.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.26.0
